Fixed an issue in PANOC init
If the Lipschitz constant was estimated to 0 at init, the solver would
get NaN as gamma, causing failure. Now gamma is estimated using the
current cost and gradient, which in limited experiments seems to be an
acceptable heuristic.

The heuristic is only applied if  the Lipschitz constant was estimated to 0.